Updated diamond: no links and smoother run

+++ diamond.sh
@@ -6,35 +6,35 @@ module diamond
 # variables and directories
 wdir="/analyses/users/nokuzothan/disc_pipe/init_tools"
 current_dir="${wdir}/diamond"
-input_reads_dir="${wdir}/megahit/output"
+input_reads_dir="${wdir}/megahit/output/default"
 input_proteins="${current_dir}/input/protein_sequences.fasta"
 output="${current_dir}/output"
 threads=$((`/bin/nproc` -2))
 
 #clear existing output directory if any, make new output directory 
-rm -rf ${output}
+if [[ -e $output ]]; then
+  rm -rf ${output} 
+fi
 mkdir -p -m a=rwx ${output}
-
-#loop through each of the files created in megahit output directory to find final.congtigs.fa files
-### Linking out Contigs
-# remove only old .fasta symlinks
-rm -f ${output}/*.fasta  
-
-for folder in $(ls -dl ${input_reads_dir}/* | grep ^d | awk '{print $9}'); do
-  file_name=`basename ${folder}`
-  ln -s ${folder}/final.contigs.fa ${output}/${file_name}.fasta
-done
-
 
 #make diamond protein database
 diamond makedb --in ${input_proteins} -d ${output}/nr
 
-#alignment using blastx
-for contigs in ${output}/*.fasta; do
-sample=$(basename ${contigs} .fasta)
-mkdir -p -m a=rwx "${output}/${sample}"
-sample_out=${output}/${sample}
+#loop through each of the files created in megahit output directory to find final.congtigs.fa files and run diamond
+for folder in ${input_reads_dir}/*; do
 
-diamond blastx -d ${output}/nr.dmnd -q ${contigs} -o $sample_out/matches.m8 --threads ${threads} -f 6
+  if [[ -d ${folder} ]]; then
+    sample=$(basename ${folder})
+    contigs=${folder}/final.contigs.fa
 
+
+    #alignment using blastx
+    if [[ -f ${contigs} ]]; then
+    sample_out=${output}/${sample}.matches.m8
+    diamond blastx -d ${output}/nr.dmnd -q ${contigs} -o ${sample_out} --threads ${threads} -f 6
+
+    else 
+      echo "Contigs file for ${sample} not found."
+    fi
+  fi
 done
